% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesSurv_AFT2.R
\name{BayesSurv_AFT2}
\alias{BayesSurv_AFT2}
\title{Bayesian Accelerated Failure Time Model}
\usage{
BayesSurv_AFT2(Y, Xmat, hyperParams, startValues_vec, mcmcParams)
}
\arguments{
\item{Y}{a matrix with three columns: left end of interval, right end of interval, left truncation time}

\item{mcmcParams}{}
}
\value{

}
\description{
Bayesian Accelerated Failure Time Model
}
