% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesSurv_AFTtv.R
\name{BayesSurv_AFTtv}
\alias{BayesSurv_AFTtv}
\title{Bayesian Accelerated Failure Time Model with Percentile-varying Effects}
\usage{
BayesSurv_AFTtv(Y, Xmat, hyperParams, startValues_vec, mcmcParams)
}
\arguments{
\item{Y}{a matrix with three columns: left end of interval, right end of interval, left truncation time}

\item{mcmcParams}{}
}
\value{

}
\description{
Bayesian Accelerated Failure Time Model with Percentile-varying Effects
}
