% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nll_AFTtv.R
\name{nll_AFTtv}
\alias{nll_AFTtv}
\title{Title}
\usage{
nll_AFTtv(
  para,
  y,
  delta,
  x_base,
  x_tv,
  baseline = "weibull",
  basis = NULL,
  tv_type,
  deg = 1,
  tstar = 1e+100,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
