% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pw_cum_mat_inv.R
\name{pw_cum_mat_inv}
\alias{pw_cum_mat_inv}
\title{Generate Matrix of Values of Inverse of Integrated Piecewise Constant Function}
\usage{
pw_cum_mat_inv(y, knots, x_base = x_base, beta_base = beta_base, x_tv, beta_tv)
}
\arguments{
\item{beta_tv}{}
}
\value{

}
\description{
Generate Matrix of Values of Inverse of Integrated Piecewise Constant Function
}
