% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_AFTtv.R
\name{sim_AFTtv}
\alias{sim_AFTtv}
\title{Simulate Data under Time-Varying AFT Model}
\usage{
sim_AFTtv(
  x_base,
  x_tv,
  beta_base_true,
  beta_tv_true,
  tv_type,
  S0_inv_func,
  cens,
  knots,
  deg = 1,
  tstar = 1e+100,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Simulate Data under Time-Varying AFT Model
}
