% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_AFTtv.R
\name{sim_AFTtv_better}
\alias{sim_AFTtv_better}
\title{Simulate Data under Time-Varying AFT Model}
\usage{
sim_AFTtv_better(
  x_base,
  x_tv,
  beta_base_true,
  beta_tv_true,
  tv_type,
  S0_inv_func = NULL,
  dist = NULL,
  intercept = NULL,
  scale = NULL,
  cens,
  knots,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Simulate Data under Time-Varying AFT Model
}
