% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_AFTtv.R
\name{sim_AFTtv_dist}
\alias{sim_AFTtv_dist}
\title{Simulate Data under Time-Varying AFT Model From Select Distributions}
\usage{
sim_AFTtv_dist(
  x_base,
  x_tv,
  beta_base_true,
  beta_tv_true,
  tv_type,
  dist,
  intercept,
  scale,
  cens,
  knots,
  deg = 1,
  tstar = 1e+100,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Simulate Data under Time-Varying AFT Model From Select Distributions
}
