% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smarg.R
\name{Smarg}
\alias{Smarg}
\title{Compute regression-standardized survival function}
\usage{
Smarg(
  t,
  x_base_aug,
  x_tv_aug,
  beta_base,
  beta_tv,
  int,
  shp,
  tv_type,
  knots = NULL,
  baseline = "weibull",
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Compute regression-standardized survival function
}
