% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smarg_inv.R
\name{Smarg_inv}
\alias{Smarg_inv}
\title{Compute inverse of regression-standardized survival function}
\usage{
Smarg_inv(
  p,
  x_base_aug,
  x_tv_aug,
  beta_base,
  beta_tv,
  int,
  shp,
  tv_type,
  knots = NULL,
  lower = 0,
  upper = 100
)
}
\arguments{
\item{upper}{}
}
\value{

}
\description{
Compute inverse of regression-standardized survival function
}
