% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sx_inv.R
\name{Sx_inv}
\alias{Sx_inv}
\title{Compute Inverse Covariate-Adjusted Survival Curve}
\usage{
Sx_inv(
  para,
  p_obj,
  x_tv = NULL,
  beta_tv = NULL,
  x_base = NULL,
  beta_base = NULL,
  baseline = "weibull",
  tv_type,
  inv_basis = NULL,
  knots = NULL,
  accelfactor = TRUE,
  logSinv = FALSE
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Compute Inverse Covariate-Adjusted Survival Curve
}
