% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vx.R
\name{Vx}
\alias{Vx}
\title{Compute Integrated Covariate Process V}
\usage{
Vx(
  t_obj,
  x_base = NULL,
  beta_base = NULL,
  x_tv = NULL,
  beta_tv = NULL,
  xbeta_base = NULL,
  xbeta_tv = NULL,
  tv_type,
  basis = NULL,
  knots = NULL
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
This function computes the integrated covariate process V used for time-varying AFT models.
}
