% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vx_inv.R
\name{Vx_inv}
\alias{Vx_inv}
\title{Compute Inverse of Integrated Covariate Process V}
\usage{
Vx_inv(
  t_obj,
  x_tv = NULL,
  beta_tv = NULL,
  xbeta_base = NULL,
  x_base = NULL,
  beta_base = NULL,
  tv_type,
  inv_basis = NULL,
  knots = NULL,
  deg = 1,
  tstar = 1e+100,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Compute Inverse of Integrated Covariate Process V
}
