% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesSurv_AFTtv.R
\name{BayesSurv_AFTtv}
\alias{BayesSurv_AFTtv}
\title{Bayesian Accelerated Failure Time Model with Percentile-varying Effects}
\usage{
BayesSurv_AFTtv(
  Y,
  Xmat,
  Xvec_tv,
  hyperParams,
  knots,
  startValues_vec,
  mcmcParams
)
}
\arguments{
\item{Y}{an n by 3 matrix with columns: left end of interval, right end of interval, left truncation time}

\item{Xmat}{an n by p matrix of covariates}

\item{Xvec_tv}{an n-length vector with the covariate having a time-varying effect}

\item{hyperParams}{a list of hyperparameters (based on SemiCompRisks for now)}

\item{mcmcParams}{a list of mcmc tuning parameters (based on SemiCompRisks for now)}

\item{startValues}{a vector of starting values for the parameters}
}
\value{
a list with outputs
}
\description{
Bayesian Accelerated Failure Time Model with Percentile-varying Effects
}
